ALUMNOS/MDAB/ALEJANDRO_BOSCH/GCP_ALMACENAMIENTO/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.10
google-auth==2.37.0
google-cloud-pubsub==2.27.1
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
pandas==2.2.3
pyarrow==19.0.0